+++ conan/tools/gnu/__init__.py
@@ -1,5 +1,6 @@
 from conan.tools.gnu.autotools import Autotools
-from conan.tools.gnu.autotoolstoolchain import AutotoolsToolchain, AutotoolsToolchainX
+from conan.tools.gnu.autotoolstoolchain import AutotoolsToolchain
+from conan.tools.gnu.gnutoolchain import GnuToolchain
 from conan.tools.gnu.autotoolsdeps import AutotoolsDeps
 from conan.tools.gnu.pkgconfig import PkgConfig
 from conan.tools.gnu.pkgconfigdeps import PkgConfigDeps
