+++ conans/client/generators/__init__.py
@@ -21,7 +21,7 @@ _generators = {"CMakeToolchain": "conan.tools.cmake",
                "VirtualBuildEnv": "conan.tools.env.virtualbuildenv",
                "AutotoolsDeps": "conan.tools.gnu",
                "AutotoolsToolchain": "conan.tools.gnu",
-               "AutotoolsToolchainX": "conan.tools.gnu",
+               "GnuToolchain": "conan.tools.gnu",
                "PkgConfigDeps": "conan.tools.gnu",
                "BazelDeps": "conan.tools.google",
                "BazelToolchain": "conan.tools.google",
